Make dashboard item one component
